Oppdatert regler og splittet ut oppdatering og regelvalidering og regelevaluering
