--- src/main/kotlin/no/nav/familie/ef/sak/regler/Regel.kt
@@ -1,14 +1,0 @@
-package no.nav.familie.ef.sak.regler
-
-import no.nav.familie.ef.sak.repository.domain.Vilkårsresultat
-
-enum class BegrunnelseType {
-    UTEN,
-    PÅKREVD,
-    VALGFRI
-}
-
-enum class Resultat(val vilkårsresultat: Vilkårsresultat) {
-    OPPFYLT(Vilkårsresultat.OPPFYLT),
-    IKKE_OPPFYLT(Vilkårsresultat.IKKE_OPPFYLT)
-}
