+++ src/main/kotlin/no/nav/familie/ef/sak/regler/vilkårsregel/AleneomsorgRegel.kt
@@ -2,10 +2,10 @@ package no.nav.familie.ef.sak.regler.vilkårsregel
 
 import no.nav.familie.ef.sak.regler.RegelId
 import no.nav.familie.ef.sak.regler.RegelSteg
-import no.nav.familie.ef.sak.regler.SluttRegel
+import no.nav.familie.ef.sak.regler.SluttSvarRegel
 import no.nav.familie.ef.sak.regler.SvarId
 import no.nav.familie.ef.sak.regler.Vilkårsregel
-import no.nav.familie.ef.sak.regler.jaNeiMapping
+import no.nav.familie.ef.sak.regler.jaNeiSvarRegel
 import no.nav.familie.ef.sak.regler.regelIds
 import no.nav.familie.ef.sak.repository.domain.VilkårType
 
@@ -21,8 +21,8 @@ class AleneomsorgRegel : Vilkårsregel(vilkårType = VilkårType.ALENEOMSORG,
 
         private val MER_AV_DAGLIG_OMSORG =
                 RegelSteg(regelId = RegelId.MER_AV_DAGLIG_OMSORG,
-                          svarMapping = jaNeiMapping(hvisJa = SluttRegel.OPPFYLT_MED_PÅKREVD_BEGRUNNELSE,
-                                                     hvisNei = SluttRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE))
+                          svarMapping = jaNeiSvarRegel(hvisJa = SluttSvarRegel.OPPFYLT_MED_PÅKREVD_BEGRUNNELSE,
+                                                       hvisNei = SluttSvarRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE))
 
         private val næreBoForholdMapping =
                 setOf(SvarId.SAMME_HUS_OG_FÆRRE_ENN_4_BOENHETER,
@@ -31,16 +31,16 @@ class AleneomsorgRegel : Vilkårsregel(vilkårType = VilkårType.ALENEOMSORG,
                       SvarId.SELVSTENDIGE_BOLIGER_SAMME_TOMT,
                       SvarId.NÆRMESTE_BOLIG_ELLER_REKKEHUS_I_SAMMEGATE,
                       SvarId.TILSTØTENDE_BOLIGER_ELLER_REKKEHUS_I_SAMMEGATE)
-                        .map { it to SluttRegel.OPPFYLT_MED_PÅKREVD_BEGRUNNELSE }
-                        .toMap() + mapOf(SvarId.NEI to SluttRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE)
+                        .map { it to SluttSvarRegel.OPPFYLT_MED_PÅKREVD_BEGRUNNELSE }
+                        .toMap() + mapOf(SvarId.NEI to SluttSvarRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE)
         private val NÆRE_BOFORHOLD =
                 RegelSteg(regelId = RegelId.NÆRE_BOFORHOLD,
                           svarMapping = næreBoForholdMapping)
 
         private val SKRIFTLIG_AVTALE_OM_DELT_BOSTED =
                 RegelSteg(regelId = RegelId.SKRIFTLIG_AVTALE_OM_DELT_BOSTED,
-                          jaNeiMapping(hvisJa = SluttRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE,
-                                       hvisNei = SluttRegel.OPPFYLT_MED_VALGFRI_BEGRUNNELSE))
+                          jaNeiSvarRegel(hvisJa = SluttSvarRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE,
+                                         hvisNei = SluttSvarRegel.OPPFYLT_MED_VALGFRI_BEGRUNNELSE))
 
     }
 }
