+++ src/main/kotlin/no/nav/familie/ef/sak/regler/vilkårsregel/ForutgåendeMedlemskapRegel.kt
@@ -3,10 +3,10 @@ package no.nav.familie.ef.sak.regler.vilkårsregel
 import no.nav.familie.ef.sak.regler.NesteRegel
 import no.nav.familie.ef.sak.regler.RegelId
 import no.nav.familie.ef.sak.regler.RegelSteg
-import no.nav.familie.ef.sak.regler.SluttRegel
+import no.nav.familie.ef.sak.regler.SluttSvarRegel
 import no.nav.familie.ef.sak.regler.SvarId
 import no.nav.familie.ef.sak.regler.Vilkårsregel
-import no.nav.familie.ef.sak.regler.jaNeiMapping
+import no.nav.familie.ef.sak.regler.jaNeiSvarRegel
 import no.nav.familie.ef.sak.regler.regelIds
 import no.nav.familie.ef.sak.repository.domain.VilkårType
 
@@ -24,16 +24,16 @@ class ForutgåendeMedlemskapRegel : Vilkårsregel(vilkårType = VilkårType.FORU
                 SvarId.ANDRE_FORELDER_MEDLEM_MINST_5_ÅR_AVBRUDD_MINDRE_ENN_10_ÅR,
                 SvarId.ANDRE_FORELDER_MEDLEM_MINST_7_ÅR_AVBRUDD_MER_ENN_10_ÅR,
                 SvarId.TOTALVURDERING_OPPFYLLER_FORSKRIFT)
-                                                .map { it to SluttRegel.OPPFYLT_MED_PÅKREVD_BEGRUNNELSE }.toMap() +
-                                        mapOf(SvarId.NEI to SluttRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE)
+                                                .map { it to SluttSvarRegel.OPPFYLT_MED_PÅKREVD_BEGRUNNELSE }.toMap() +
+                                        mapOf(SvarId.NEI to SluttSvarRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE)
         private val MEDLEMSKAP_UNNTAK =
                 RegelSteg(regelId = RegelId.MEDLEMSKAP_UNNTAK,
                           svarMapping = unntakSvarMapping)
 
         private val SØKER_MEDLEM_I_FOLKETRYGDEN =
                 RegelSteg(regelId = RegelId.SØKER_MEDLEM_I_FOLKETRYGDEN,
-                          svarMapping = jaNeiMapping(hvisJa = SluttRegel.OPPFYLT_MED_VALGFRI_BEGRUNNELSE,
-                                                     hvisNei = NesteRegel(MEDLEMSKAP_UNNTAK.regelId)))
+                          svarMapping = jaNeiSvarRegel(hvisJa = SluttSvarRegel.OPPFYLT_MED_VALGFRI_BEGRUNNELSE,
+                                                       hvisNei = NesteRegel(MEDLEMSKAP_UNNTAK.regelId)))
     }
 
 }
