+++ src/main/kotlin/no/nav/familie/ef/sak/regler/vilkårsregel/NyttBarnSammePartnerRegel.kt
@@ -2,9 +2,9 @@ package no.nav.familie.ef.sak.regler.vilkårsregel
 
 import no.nav.familie.ef.sak.regler.RegelId
 import no.nav.familie.ef.sak.regler.RegelSteg
-import no.nav.familie.ef.sak.regler.SluttRegel
+import no.nav.familie.ef.sak.regler.SluttSvarRegel
 import no.nav.familie.ef.sak.regler.Vilkårsregel
-import no.nav.familie.ef.sak.regler.jaNeiMapping
+import no.nav.familie.ef.sak.regler.jaNeiSvarRegel
 import no.nav.familie.ef.sak.regler.regelIds
 import no.nav.familie.ef.sak.repository.domain.VilkårType
 
@@ -16,7 +16,7 @@ class NyttBarnSammePartnerRegel : Vilkårsregel(vilkårType = VilkårType.NYTT_B
 
         private val HAR_FÅTT_ELLER_VENTER_NYTT_BARN_MED_SAMME_PARTNER =
                 RegelSteg(regelId = RegelId.HAR_FÅTT_ELLER_VENTER_NYTT_BARN_MED_SAMME_PARTNER,
-                          svarMapping = jaNeiMapping(hvisJa = SluttRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE,
-                                                     hvisNei = SluttRegel.OPPFYLT_MED_VALGFRI_BEGRUNNELSE))
+                          svarMapping = jaNeiSvarRegel(hvisJa = SluttSvarRegel.IKKE_OPPFYLT_MED_PÅKREVD_BEGRUNNELSE,
+                                                       hvisNei = SluttSvarRegel.OPPFYLT_MED_VALGFRI_BEGRUNNELSE))
     }
 }
